Move Community type to models, lookup to services
