+++ src/types/models.ts
@@ -1,3 +1,8 @@
+export interface Community {
+  slug: string;
+  name: string;
+}
+
 export interface Reference {
   id: string;
 }
